Add custom DOM node identifiers which we can use to cross-reference DOM nodes in the solver.

+++ artemis-code/src/concolic/solver/constraintwriter/cvc4.cpp
@@ -439,7 +439,7 @@ void CVC4ConstraintWriter::visit(Symbolic::StringRegexSubmatchArrayMatch* exp, v
 void CVC4ConstraintWriter::visit(Symbolic::ConstantObject* obj, void* arg)
 {
     std::stringstream instanceIdentifier;
-    instanceIdentifier << (unsigned long)obj->getInstanceidentifier();
+    instanceIdentifier << obj->getInstanceidentifier();
 
     mExpressionType = Symbolic::OBJECT;
     mExpressionBuffer = instanceIdentifier.str(); // 0 if the reference is null or undefined, otherwise its a number
